requirements.txt
numpy>=1.20.3,<2.0
Flask>=2.2.2
SQLAlchemy>=2.0.20
mindspore>=2.2.14
Pillow>=8.1.0
pillow-heif>=0.14.0
transformers>=4.28.1
opencv-python-headless>=4.7.0.68
python-dotenv>=0.19.2
tqdm>=4.66.1
requests>=2.31.0
git+https://github.com/XixinYang/CLIP-mindspore.git
torch>=1.7.0